Minor fixes

+++ src/libraries/v2mp_asm/src/ProgramModel/ProgramModel.cpp
@@ -12,7 +12,7 @@ namespace V2MPAsm
 			return;
 		}
 
-		codeWord->SetAddress(m_CodeWords.size() * 2);
+		codeWord->SetAddress(static_cast<uint16_t>(m_CodeWords.size() * 2));
 		m_CodeWords.emplace_back(codeWord);
 	}
